refactor(echo-test): echotest -> echo_test
